support for wrap logging
